requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pytesseract
Pillow
pytest
httpx
opencv-python-headless
bulma 